services/preprocessing-service/requirements.txt
grpcio>=1.62.0
